Preserve Vite lazy Babel contract tests
